+++ rust/lance-index/Cargo.toml
@@ -125,5 +125,13 @@ harness = false
 name = "inverted"
 harness = false
 
+[[bench]]
+name = "kmeans"
+harness = false
+
+[[bench]]
+name = "compute_partition"
+harness = false
+
 [lints]
 workspace = true
