+++ rust/lance-index/src/vector/flat/storage.rs
@@ -105,7 +105,7 @@ impl QuantizerStorage for FlatFloatStorage {
 }
 
 impl FlatFloatStorage {
-    // deprecated, use `try_from_batch` instead
+    // used for only testing
     pub fn new(vectors: FixedSizeListArray, distance_type: DistanceType) -> Self {
         let row_ids = Arc::new(UInt64Array::from_iter_values(0..vectors.len() as u64));
         let vectors = Arc::new(vectors);
@@ -263,6 +263,28 @@ impl QuantizerStorage for FlatBinStorage {
 }
 
 impl FlatBinStorage {
+    // used for only testing
+    pub fn new(vectors: FixedSizeListArray, distance_type: DistanceType) -> Self {
+        let row_ids = Arc::new(UInt64Array::from_iter_values(0..vectors.len() as u64));
+        let vectors = Arc::new(vectors);
+
+        let batch = RecordBatch::try_from_iter_with_nullable(vec![
+            (ROW_ID, row_ids.clone() as ArrayRef, true),
+            (FLAT_COLUMN, vectors.clone() as ArrayRef, true),
+        ])
+        .unwrap();
+
+        Self {
+            metadata: FlatMetadata {
+                dim: vectors.value_length() as usize,
+            },
+            batch,
+            distance_type,
+            row_ids,
+            vectors,
+        }
+    }
+
     pub fn vector(&self, id: u32) -> ArrayRef {
         self.vectors.value(id as usize)
     }
