+++ rust/lance-linalg/Cargo.toml
@@ -70,13 +70,5 @@ harness = false
 name = "norm_l2"
 harness = false
 
-[[bench]]
-name = "kmeans"
-harness = false
-
-[[bench]]
-name = "compute_partition"
-harness = false
-
 [lints]
 workspace = true
