+++ rust/lance/src/utils.rs
@@ -12,4 +12,4 @@ pub mod tfrecord;
 
 // Re-export
 pub use lance_datafusion::sql;
-pub use lance_linalg::kmeans;
+pub use lance_index::vector::kmeans;
